Refactor `aviation` module to `fleet` module
