+++ tests/test_aviation.py
@@ -1,6 +1,6 @@
 import pytest
 
-from aviation.aviation import passengers_per_day, required_global_fleet
+from aviation.fleet import passengers_per_day, required_global_fleet
 
 
 @pytest.mark.parametrize(
